Se en esta versión se hicieron muchos cambios aquí un resumen: 1.- modificación de controladores, 2.- creación de politicas, 3.- habilitación de la autorización, 4.- accesors y mutators, 5.- ventanas de error, 6.- cambios en algunas consultas, 7.- solucioón de errores en actulizar tanto de notas como de bugs

+++ app/Http/Controllers/BugController.php
@@ -16,6 +16,9 @@ class BugController extends Controller
     public function index()
     {
         $bugs=Bug::where('user_id',auth()->user()->id)->get();
+
+        $this->authorize('viewany',Bug::class);
+
         return view('bugs.index', compact('bugs'));
     }
 
@@ -25,6 +28,9 @@ class BugController extends Controller
     public function create()
     {
         $asignatura= Subject::where('ing',auth()->user()->ing)->get();
+        
+        $this->authorize('create', Bug::class);
+
         return view('bugs.create', compact('asignatura'));
     }
 
@@ -33,13 +39,16 @@ class BugController extends Controller
      */
     public function store(Request $request)
     {
+
+        $this->authorize('create', Bug::class);
+
         $validator = Validator::make($request->all(), [
-            'description' => 'required',
-            'codigo' => 'required',
-            'solution' => 'required',
-            'plataforma'=>'required',
-            'estado' => 'required',
-            'asignatura'=>'required'
+            'description' => 'required|string|max:255',
+            'codigo' => 'required|string|max:100',
+            'solution' => 'required|string|max:255',
+            'plataforma'=>'required|string|max:100',
+            'estado' => 'required', // "integer" validacion de que sea solo numero no necesaria
+            'asignatura'=>'required' // "integer" validacion de que sea solo numero no necesaria
         ]);
 
         if ($validator->fails()) {
@@ -59,13 +68,15 @@ class BugController extends Controller
         $bug->save();
         return redirect()->route('bugs.index');
     }
-
     /**
      * Display the specified resource.
      */
     public function show($id)
     {
         $detalle_bug=Bug::find($id);  
+
+        $this->authorize('view', $detalle_bug);
+
         return view('bugs.show', compact('detalle_bug'));
     }
 
@@ -75,6 +86,9 @@ class BugController extends Controller
     public function edit($id)
     {
         $bug=Bug::find($id);
+
+        $this->authorize('update', $bug);
+
         return view('bugs.edit',compact('bug'));
     }
 
@@ -95,16 +109,14 @@ class BugController extends Controller
                         ->withInput();
         }
         
-        $dbug=Bug::find($id);
-
         $bug = Bug::find($id);
-        $bug->description=$request->description;
-        $bug->codigo=$dbug->codigo;
-        $bug->solution=$request->solution;
-        $bug->plataforma=$dbug->plataforma;
-        $bug->estado=$request->estado;
+
+        $this->authorize('update', $bug);
+
+        $bug->description = $request->description;
+        $bug->solution = $request->solution;
+        $bug->estado = $request->estado;
         $bug->user_id = auth()->user()->id;
-        $bug->subject_id=$dbug->subject_id;
         $bug->save();
         return redirect()->route('bugs.index');
     }
@@ -115,6 +127,9 @@ class BugController extends Controller
     public function destroy($id)
     {
         $bug=Bug::find($id);
+
+        $this->authorize('delete', $bug);
+
         $bug->delete();
         return redirect()->route('bugs.index');
     }
